+++ CHANGELOG.md
@@ -4,3 +4,5 @@
    * begin dusting off this old package
    * move tests into subdirectory, prepare for multiple test packages
    * use shields io npm badge
+   * add tests for export patterns from nodejs documentation
+   * add exports.import definition